Using Arrays convert unstable selection sort as stable selection sort
A sorting algorithm is said to be stable if two objects with equal or same keys appear in the same order in sorted output as they appear in the input array to be sorted.
Any comparison based sorting algorithm which is not stable by nature can be modified to be stable by changing the key comparison operation so that the comparison of two keys considers position as a factor for objects with equal key or by tweaking it in a way such that its meaning doesn’t change and it becomes stable as well.

Selection sort works by finding the minimum element and then inserting it in its correct position by swapping with the element which is in the position of this minimum element. This is what makes it unstable.

Swapping might impact in pushing a key (imagine X)to a position greater than the key(Imagine Y) which are equal keys. which makes them out of desired order.

Let us understand using below example:
Example: 
            this is give array let us mark X and Y for similar elements
             4X 5 3 2 4Y 1
         First minimum element is 1, now instead
         of swapping. Insert 1 in its correct place 
         and pushing every element one step forward
         i.e forward pushing.
         4X 5 3 2 4Y
         Next minimum is 2 :
         1 2 4X 5 3 4Y
         Next minimum is 3 :
         1 2 3 4X 5 4Y
         Repeat the steps until array is sorted.
         1 2 3 4X 4Y 5