Create ConstructiveAlgo.java
Andy wants to play a game with his little brother, Bob. The game starts with an array of distinct integers and the rules are as follows:

-->Bob always plays first and the two players move in alternating turns.
-->In a single move, a player chooses the maximum element currently present in the array and removes it as well as all the other elements to its right. For example, if the starting array arr=[2,3,5,4,1] then it becomes arr`=[2,3]  after the first move because we remove the maximum element (i.e., 5) and all elements to its right. (i.e., 4 and  1).
-->The modifications made to the array during each turn are permanent, so the next player continues the game with the remaining array. The first player who is unable to make a move loses the game.
Andy and Bob play g games. Given the initial array for each game, find and print the name of the winner on a new line. If Andy wins, print ANDY; if Bob wins, print BOB.

To continue the example above, in the next move Andy will remove . Bob 3 will then remove 2 and win because there are no more integers to remove.